+++ lib/polar.dart
@@ -3,11 +3,10 @@ library polar;
 export 'src/polar_base.dart';
 export 'src/events.dart';
 
-export 'src/model/device_streaming_feature.dart' show DeviceStreamingFeature;
-export 'src/model/ohr_data_type.dart' show OhrDataType;
+export 'src/model/polar_ble_sdk_feature.dart';
+export 'src/model/polar_device_data_type.dart';
+export 'src/model/ppg_data_type.dart';
 export 'src/model/polar_device_info.dart';
-export 'src/model/polar_hr_data.dart';
 export 'src/model/polar_recording.dart';
-export 'src/model/polar_sensor_setting.dart'
-    show PolarSensorSetting, PolarSettingType;
+export 'src/model/polar_sensor_setting.dart';
 export 'src/model/polar_streaming.dart';