+++ lib/src/model/polar_sensor_setting.g.dart
@@ -11,14 +11,22 @@ part of 'polar_sensor_setting.dart';
 PolarSensorSetting _$PolarSensorSettingFromJson(Map<String, dynamic> json) =>
     PolarSensorSetting._(
       settings: (json['settings'] as Map<String, dynamic>).map(
-        (k, e) => MapEntry(
-            const PolarSettingTypeConverter().fromJson(k as String),
+        (k, e) => MapEntry($enumDecode(_$PolarSettingTypeEnumMap, k),
             (e as List<dynamic>).map((e) => e as int).toList()),
       ),
     );
 
 Map<String, dynamic> _$PolarSensorSettingToJson(PolarSensorSetting instance) =>
     <String, dynamic>{
-      'settings': instance.settings.map(
-          (k, e) => MapEntry(const PolarSettingTypeConverter().toJson(k), e)),
+      'settings': instance.settings
+          .map((k, e) => MapEntry(_$PolarSettingTypeEnumMap[k]!, e)),
     };
+
+const _$PolarSettingTypeEnumMap = {
+  PolarSettingType.sampleRate: 'sampleRate',
+  PolarSettingType.resolution: 'resolution',
+  PolarSettingType.range: 'range',
+  PolarSettingType.rangeMilliunit: 'rangeMilliunit',
+  PolarSettingType.channels: 'channels',
+  PolarSettingType.unknown: 'unknown',
+};