+++ lib/src/model/polar_streaming.g.dart
@@ -8,16 +8,30 @@ part of 'polar_streaming.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
+PolarStreamingData<T> _$PolarStreamingDataFromJson<T>(
+  Map<String, dynamic> json,
+  T Function(Object? json) fromJsonT,
+) =>
+    PolarStreamingData<T>(
+      samples: (json['samples'] as List<dynamic>).map(fromJsonT).toList(),
+    );
+
+PolarHrSample _$PolarHrSampleFromJson(Map<String, dynamic> json) =>
+    PolarHrSample(
+      timeStamp: json['timeStamp'] as int,
+      hr: json['hr'] as int,
+      rrs: (json['rrs'] as List<dynamic>).map((e) => e as int).toList(),
+      rrsMs: (json['rrsMs'] as List<dynamic>).map((e) => e as int).toList(),
+      contactStatus:
+          PolarHrSample._readContactStatus(json, 'contactStatus') as bool,
+      contactStatusSupported: PolarHrSample._readContactStatusSupported(
+          json, 'contactStatusSupported') as bool,
+    );
+
 PolarEcgSample _$PolarEcgSampleFromJson(Map<String, dynamic> json) =>
     PolarEcgSample(
       timeStamp: json['timeStamp'] as int,
       voltage: json['voltage'] as int,
-    );
-
-PolarEcgData _$PolarEcgDataFromJson(Map<String, dynamic> json) => PolarEcgData(
-      samples: (json['samples'] as List<dynamic>)
-          .map((e) => PolarEcgSample.fromJson(e as Map<String, dynamic>))
-          .toList(),
     );
 
 PolarAccSample _$PolarAccSampleFromJson(Map<String, dynamic> json) =>
@@ -28,25 +42,12 @@ PolarAccSample _$PolarAccSampleFromJson(Map<String, dynamic> json) =>
       z: json['z'] as int,
     );
 
-PolarAccData _$PolarAccDataFromJson(Map<String, dynamic> json) => PolarAccData(
-      samples: (json['samples'] as List<dynamic>)
-          .map((e) => PolarAccSample.fromJson(e as Map<String, dynamic>))
-          .toList(),
-    );
-
 PolarGyroSample _$PolarGyroSampleFromJson(Map<String, dynamic> json) =>
     PolarGyroSample(
       timeStamp: json['timeStamp'] as int,
       x: (json['x'] as num).toDouble(),
       y: (json['y'] as num).toDouble(),
       z: (json['z'] as num).toDouble(),
-    );
-
-PolarGyroData _$PolarGyroDataFromJson(Map<String, dynamic> json) =>
-    PolarGyroData(
-      samples: (json['samples'] as List<dynamic>)
-          .map((e) => PolarGyroSample.fromJson(e as Map<String, dynamic>))
-          .toList(),
     );
 
 PolarMagnetometerSample _$PolarMagnetometerSampleFromJson(
@@ -58,46 +59,30 @@ PolarMagnetometerSample _$PolarMagnetometerSampleFromJson(
       z: (json['z'] as num).toDouble(),
     );
 
-PolarMagnetometerData _$PolarMagnetometerDataFromJson(
-        Map<String, dynamic> json) =>
-    PolarMagnetometerData(
-      samples: (json['samples'] as List<dynamic>)
-          .map((e) =>
-              PolarMagnetometerSample.fromJson(e as Map<String, dynamic>))
-          .toList(),
-    );
-
-PolarOhrSample _$PolarOhrSampleFromJson(Map<String, dynamic> json) =>
-    PolarOhrSample(
+PolarPpgSample _$PolarPpgSampleFromJson(Map<String, dynamic> json) =>
+    PolarPpgSample(
       timeStamp: json['timeStamp'] as int,
       channelSamples: (json['channelSamples'] as List<dynamic>)
           .map((e) => e as int)
           .toList(),
     );
 
-PolarOhrData _$PolarOhrDataFromJson(Map<String, dynamic> json) => PolarOhrData(
-      type: const OhrDataTypeConverter().fromJson(json['type']),
+PolarPpgData _$PolarPpgDataFromJson(Map<String, dynamic> json) => PolarPpgData(
+      type: const PpgDataTypeConverter().fromJson(json['type']),
       samples: (json['samples'] as List<dynamic>)
-          .map((e) => PolarOhrSample.fromJson(e as Map<String, dynamic>))
+          .map((e) => PolarPpgSample.fromJson(e as Map<String, dynamic>))
           .toList(),
     );
 
-PolarOhrPpiSample _$PolarOhrPpiSampleFromJson(Map<String, dynamic> json) =>
-    PolarOhrPpiSample(
-      ppi: PolarOhrPpiSample._readPpi(json, 'ppi') as int,
+PolarPpiSample _$PolarPpiSampleFromJson(Map<String, dynamic> json) =>
+    PolarPpiSample(
+      ppi: PolarPpiSample._readPpi(json, 'ppi') as int,
       errorEstimate:
-          PolarOhrPpiSample._readErrorEstimate(json, 'errorEstimate') as int,
+          PolarPpiSample._readErrorEstimate(json, 'errorEstimate') as int,
       hr: json['hr'] as int,
       blockerBit: const PlatformBooleanConverter().fromJson(json['blockerBit']),
       skinContactStatus:
           const PlatformBooleanConverter().fromJson(json['skinContactStatus']),
       skinContactSupported: const PlatformBooleanConverter()
           .fromJson(json['skinContactSupported']),
-    );
-
-PolarOhrPpiData _$PolarOhrPpiDataFromJson(Map<String, dynamic> json) =>
-    PolarOhrPpiData(
-      samples: (json['samples'] as List<dynamic>)
-          .map((e) => PolarOhrPpiSample.fromJson(e as Map<String, dynamic>))
-          .toList(),
     );